requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
mysql-connector-python==9.2.0
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.6.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
